requirements.txt
openai-whisper>=20231117
torch>=2.1.0
requests>=2.31.0
gradio>=4.44.0
